requirements.txt
torch==1.3.1
torchvision==0.4.2
numpy==1.17.4
opencv-python==4.2.0.32
pillow==6.2.1
tqdm==4.43.0
munkres==1.1.2
cython==0.29.15
pycocotools==2.0.0
scikit-image==0.15.0
scipy==1.3.0
matplotlib==3.0.0
pandas==0.25.1
cvbase==0.5.5